Remove maskElements array and remove all ShimmerContainer children on setLayout
MaskElements was an array that was seemingly unused, but multiple calls to setLayout would retain all previously held views and append the new views as well.

ShimmerContainer would eventually grow in size as multiple setLayout calls were made. This can be remedied by removing all the views contained on setLayout.